services/calendar/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Cors==5.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.32.3
pika==1.3.2 